services/content_service/pyproject.toml
[tool.poetry]
name = "content-service"
version = "0.1.0"
description = ""
authors = ["Yury Zimarev <yury.zimarev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.11"
sqlalchemy = "^2.0.38"
pydantic = {extras = ["email"], version = "^2.10.6"}
python-jose = "^3.4.0"
pydantic-settings = "^2.8.1"
uvicorn = "^0.34.0"
psycopg2-binary = "^2.9.10"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
